requirements.txt
Jinja2~=3.1
Markdown~=3.5
PyYAML~=6.0
